+++ Reddit/Easy/closest_string/src/lib.rs
@@ -12,7 +12,6 @@ pub fn closest_string(sequences: &Vec<String>) -> String {
     let mut smallest_sum: usize = sequences.clone().iter()
         .map(|s| hamming_distance(smallest_seq, s)).sum();
 
-    // let mut iter = sequences.iter();
     // TODO: space for performance improvement
     for one in sequences.iter().skip(1) {
         let sum = sequences.clone().iter()
@@ -22,8 +21,6 @@ pub fn closest_string(sequences: &Vec<String>) -> String {
             smallest_sum = sum;
             smallest_seq = one;
         }
-
-        // iter.next();
     }
 
     smallest_seq.to_string()