+++ Reddit/Easy/integer_complexity_1/src/lib.rs
@@ -1,6 +1,8 @@
 
-pub fn stub() {
-    println!("hello there");
+pub fn smallest_factor_sum(n: i32) -> i32 {
+    println!("got: {}", n);
+
+    0
 }
 
 #[cfg(test)]